Implement preemptive ratelimiting
The implementation is very simple; once the X-RateLimit-Remaining counter has reached zero, it waits the appropriate amount of seconds, which is determined by taking the difference between the HTTP Date header (representing the "now") and the specified reset time (representing the "then").

As `sync_wait` is called with the specific method mutex, no other thread will be able to use that endpoint until the time has run out.